[DAT-6609] execute-sql refactoring
