+++ liquibase-standard/src/main/resources/META-INF/services/liquibase.command.LiquibaseCommand
@@ -1,7 +1,6 @@
 liquibase.command.core.DiffCommand
 liquibase.command.core.DiffToChangeLogCommand
 liquibase.command.core.DropAllCommand
-liquibase.command.core.ExecuteSqlCommand
 liquibase.command.core.GenerateChangeLogCommand
 liquibase.command.core.HistoryCommand
 liquibase.command.core.SnapshotCommand
